include lint hints
