--- frontend/ctfl_vertragsmanager/lib/constants/Color_Themes.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ColorThemes {
-  static Color _primaryColor = Color(0xffD72324);
-  static Color get primaryColor => _primaryColor;
-}
