+++ frontend/ctfl_vertragsmanager/lib/constants/colors.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+class ColorThemes {
+  static const Color _primaryColor = Color(0xffD72324);
+  static Color get primaryColor => _primaryColor;
+}
