--- frontend/ctfl_vertragsmanager/lib/funktionen/imageUtils.dart
@@ -1,15 +1,0 @@
-import 'dart:async';
-import 'dart:io';
-import 'package:flutter/services.dart';
-import 'package:path_provider/path_provider.dart';
-
-class ImageUtils {
-  static Future<File> imageToFile({String? imageName, String? ext}) async {
-    var bytes = await rootBundle.load('assets/$imageName.$ext');
-    String tempPath = (await getTemporaryDirectory()).path;
-    File file = File('$tempPath/profile.png');
-    await file.writeAsBytes(
-        bytes.buffer.asUint8List(bytes.offsetInBytes, bytes.lengthInBytes));
-    return file;
-  }
-}
