+++ frontend/ctfl_vertragsmanager/lib/models/profile.dart
@@ -1,38 +1,20 @@
 class Profile {
-  String? _id;
-  String? get id => this._id;
+  String? id;
 
-  set id(String? value) => this._id = value;
+  String email;
 
-  String _email;
-  String get email => this._email;
+  String password;
 
-  set email(String value) => this._email = value;
+  String accessToken;
 
-  String _password;
-  String get password => this._password;
-
-  set password(String value) => this._password = value;
-
-  String _accessToken;
-  String get accessToken => this._accessToken;
-
-  set accessToken(String value) => this._accessToken = value;
-
-  String _refreshToken;
-  String get refreshToken => this._refreshToken;
-
-  set refreshToken(String value) => this._refreshToken = value;
+  String refreshToken;
 
   Profile({
-    required email,
-    id,
-    password,
+    required this.email,
+    required this.password,
+    this.id,
     accessToken,
     refreshToken,
-  })  : _id = id,
-        _email = email,
-        _password = password,
-        _accessToken = accessToken ?? "",
-        _refreshToken = refreshToken ?? "";
+  })  : accessToken = accessToken ?? "",
+        refreshToken = refreshToken ?? "";
 }
