+++ frontend/ctfl_vertragsmanager/lib/pages/profil.dart
@@ -1,17 +1,14 @@
-import 'dart:convert';
 import 'dart:io';
-import 'package:ctfl_vertragsmanager/constants/Color_Themes.dart';
-import 'package:ctfl_vertragsmanager/funktionen/dbFunctions.dart';
-import 'package:ctfl_vertragsmanager/funktionen/profilFunktionen.dart';
+import 'package:ctfl_vertragsmanager/funktionen/db_functions.dart';
+import 'package:ctfl_vertragsmanager/funktionen/profil_funktionen.dart';
 import 'package:ctfl_vertragsmanager/models/profile.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class ProfilPage extends StatefulWidget {
-  late Profile user = Profile(email: "", password: "");
-  File? profilePicture = null;
+  Profile user = Profile(email: "", password: "");
+  final File? profilePicture;
+
+  ProfilPage({Key? key, this.profilePicture}) : super(key: key);
 
   @override
   State<ProfilPage> createState() => _ProfilPageState();
@@ -27,48 +24,49 @@ class _ProfilPageState extends State<ProfilPage> {
 
   @override
   void initState() {
+    super.initState();
     getUserInformation();
   }
 
-  Future _createDialogChangeProfilePicture(BuildContext context) async {
-    return showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return SimpleDialog(
-            title: const Text('Bild auswählen'),
-            children: <Widget>[
-              SimpleDialogOption(
-                onPressed: () async {
-                  _imgFromGalleryOrCamera("camera");
-                },
-                child: const Text('Kamera'),
-              ),
-              SimpleDialogOption(
-                onPressed: () async {
-                  _imgFromGalleryOrCamera("Gallery");
-                },
-                child: const Text('Galerie'),
-              ),
-            ],
-          );
-        });
-  }
+  // Future _createDialogChangeProfilePicture(BuildContext context) async {
+  //   return showDialog(
+  //       context: context,
+  //       builder: (BuildContext context) {
+  //         return SimpleDialog(
+  //           title: const Text('Bild auswählen'),
+  //           children: <Widget>[
+  //             SimpleDialogOption(
+  //               onPressed: () async {
+  //                 _imgFromGalleryOrCamera("camera");
+  //               },
+  //               child: const Text('Kamera'),
+  //             ),
+  //             SimpleDialogOption(
+  //               onPressed: () async {
+  //                 _imgFromGalleryOrCamera("Gallery");
+  //               },
+  //               child: const Text('Galerie'),
+  //             ),
+  //           ],
+  //         );
+  //       });
+  // }
 
-  Future _imgFromGalleryOrCamera(String location) async {
-    ImageSource source;
-    if (location == "Gallery")
-      source = ImageSource.gallery;
-    else
-      source = ImageSource.camera;
-    try {
-      final image = await ImagePicker().pickImage(source: source, imageQuality: 50);
-      if (image == null) return;
-      final imageTemp = File(image.path);
-      Navigator.of(context).pop();
-    } on PlatformException catch (e) {
-      print(e);
-    }
-  }
+  // Future _imgFromGalleryOrCamera(String location) async {
+  //   ImageSource source;
+  //   if (location == "Gallery")
+  //     source = ImageSource.gallery;
+  //   else
+  //     source = ImageSource.camera;
+  //   try {
+  //     final image = await ImagePicker().pickImage(source: source, imageQuality: 50);
+  //     if (image == null) return;
+  //     final imageTemp = File(image.path);
+  //     Navigator.of(context).pop();
+  //   } on PlatformException catch (e) {
+  //     print(e);
+  //   }
+  // }
 
   @override
   Widget build(BuildContext context) {
@@ -96,31 +94,31 @@ class _ProfilPageState extends State<ProfilPage> {
                 )
               ],
             ),
-            SizedBox(
+            const SizedBox(
               height: 35,
             ),
             Column(
               children: [
                 ListTile(
-                  leading: Icon(Icons.password),
-                  subtitle: Text("Möchten Sie Ihr Passwort ändern?"),
-                  title: Text("Passwort ändern"),
+                  leading: const Icon(Icons.password),
+                  subtitle: const Text("Möchten Sie Ihr Passwort ändern?"),
+                  title: const Text("Passwort ändern"),
                   onTap: () {
                     createAlertDialogChangePassword(context);
                   },
                 ),
                 ListTile(
-                  leading: Icon(Icons.login_outlined),
-                  subtitle: Text("Möchten Sie sich abmelden?"),
-                  title: Text("Abmelden"),
+                  leading: const Icon(Icons.login_outlined),
+                  subtitle: const Text("Möchten Sie sich abmelden?"),
+                  title: const Text("Abmelden"),
                   onTap: () {
                     logout(context);
                   },
                 ),
                 ListTile(
-                  leading: Icon(Icons.delete),
-                  subtitle: Text("Möchten Sie Ihr Konto unwiderruflich löschen?"),
-                  title: Text("Konto löschen"),
+                  leading: const Icon(Icons.delete),
+                  subtitle: const Text("Möchten Sie Ihr Konto unwiderruflich löschen?"),
+                  title: const Text("Konto löschen"),
                   onTap: () {
                     createAlertDialogDeleteProfile(context);
                   },
