+++ frontend/ctfl_vertragsmanager/lib/pages/vertragsdetails.dart
@@ -1,18 +1,15 @@
-import 'package:ctfl_vertragsmanager/constants/Color_Themes.dart';
-import 'package:ctfl_vertragsmanager/funktionen/dbFunctions.dart';
-import 'package:ctfl_vertragsmanager/funktionen/hiveFunctions.dart';
+import 'package:ctfl_vertragsmanager/constants/colors.dart';
+import 'package:ctfl_vertragsmanager/funktionen/db_functions.dart';
+import 'package:ctfl_vertragsmanager/funktionen/hive_functions.dart';
 import 'package:ctfl_vertragsmanager/models/vertrag.dart';
-import 'package:ctfl_vertragsmanager/models/vertragsdaten.dart';
-import 'package:ctfl_vertragsmanager/partials/customLabeledText.dart';
+import 'package:ctfl_vertragsmanager/partials/custom_labeled_text.dart';
 import 'package:ctfl_vertragsmanager/provider/all_vertraege_provider.dart';
 import 'package:ctfl_vertragsmanager/provider/cur_vertrag_provider.dart';
 import 'package:flutter/material.dart';
-import 'package:hive/hive.dart';
-import 'package:hive_flutter/hive_flutter.dart';
-import 'package:provider/src/provider.dart';
+import 'package:provider/provider.dart';
 
 class VertragsDetailsPage extends StatefulWidget {
-  VertragsDetailsPage({Key? key}) : super(key: key);
+  const VertragsDetailsPage({Key? key}) : super(key: key);
 
   @override
   State<VertragsDetailsPage> createState() => _VertragsDetailsPageState();
@@ -28,8 +25,8 @@ class _VertragsDetailsPageState extends State<VertragsDetailsPage> {
   final vertragsBox = HiveFunctions.getHiveVertraege();
 
   getVertrag() async {
-    vertrag = await context.read<cur_Vertrag_Provider>().get_cur_Vertrag();
-    context.read<cur_Vertrag_Provider>().set_cur_Vertrag_id(vertrag.id!);
+    vertrag = await context.read<CurVertragProvider>().getCurVertrag();
+    context.read<CurVertragProvider>().setCurVertragId(vertrag.id!);
     setState(() {
       loading = false;
     });
@@ -43,7 +40,7 @@ class _VertragsDetailsPageState extends State<VertragsDetailsPage> {
 
   @override
   Widget build(BuildContext context) {
-    if (loading) return Text("Loading");
+    if (loading) return const Text("Loading");
     labelColor = vertrag.label == null || vertrag.label!.colorValue == Colors.white.value
         ? ColorThemes.primaryColor
         : Color(vertrag.label!.colorValue);
@@ -52,25 +49,25 @@ class _VertragsDetailsPageState extends State<VertragsDetailsPage> {
       appBar: AppBar(
         leading: IconButton(
             onPressed: () {
-              context.read<cur_Vertrag_Provider>().reset_cur_Vertrag_id();
+              context.read<CurVertragProvider>().resetCurVertragId();
               Navigator.popAndPushNamed(context, '/main');
             },
-            icon: Icon(Icons.arrow_back)),
+            icon: const Icon(Icons.arrow_back)),
         centerTitle: true,
         backgroundColor: labelColor,
         title: Text(
           vertrag.name,
-          style: TextStyle(fontWeight: FontWeight.bold),
+          style: const TextStyle(fontWeight: FontWeight.bold),
         ),
         actions: [
           IconButton(
             onPressed: () {
               deleteVertrag(vertrag.id!);
-              context.read<cur_Vertrag_Provider>().reset_cur_Vertrag_id();
-              context.read<all_Vertraege_Provider>().notifyOurListeners();
+              context.read<CurVertragProvider>().resetCurVertragId();
+              context.read<AllVertraegeProvider>().notifyOurListeners();
               Navigator.popAndPushNamed(context, '/main');
             },
-            icon: Icon(
+            icon: const Icon(
               Icons.delete_outlined,
               size: 30,
             ),
@@ -79,20 +76,20 @@ class _VertragsDetailsPageState extends State<VertragsDetailsPage> {
       ),
       body: ListView(
         children: [
-          SizedBox(height: 15),
+          const SizedBox(height: 15),
           DetailsTile(value: vertrag.name, description: "Name", lineColor: labelColor),
-          if (vertrag.beschreibung != null && vertrag.beschreibung!.trim().length > 0)
+          if (vertrag.beschreibung != null && vertrag.beschreibung!.trim().isNotEmpty)
             DetailsTile(
                 value: vertrag.beschreibung!, description: "Beschreibung", lineColor: labelColor),
-          if (vertrag.label != null && vertrag.label!.name.trim().length > 0)
+          if (vertrag.label != null && vertrag.label!.name.trim().isNotEmpty)
             DetailsTile(
                 value: vertrag.getLabelName()!, description: "Label", lineColor: labelColor),
-          SizedBox(height: 20),
+          const SizedBox(height: 20),
           if (vertrag.intervall != null ||
               vertrag.getBeitragNumber() != null ||
               vertrag.getErstzahlung() != null ||
               vertrag.getNaechsteZahlung() != null)
-            Text(
+            const Text(
               "Zahlungsinformationen",
               textAlign: TextAlign.center,
               style: TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
@@ -101,24 +98,24 @@ class _VertragsDetailsPageState extends State<VertragsDetailsPage> {
             DetailsTile(value: vertrag.intervall!, description: "Intervall", lineColor: labelColor),
           if (vertrag.getBeitragNumber() != null)
             DetailsTile(
-                value: vertrag.getBeitragEuro()!, description: "Beitrag", lineColor: labelColor),
+                value: vertrag.getBeitragEuro(), description: "Beitrag", lineColor: labelColor),
           if (vertrag.getErstzahlung() != null)
             DetailsTile(
                 value: vertrag.getErstzahlung()!,
                 description: "Erstzahlung",
                 lineColor: labelColor),
           if (vertrag.getNaechsteZahlung() != null &&
-              vertrag.getNaechsteZahlung()!.trim().length > 0)
+              vertrag.getNaechsteZahlung()!.trim().isNotEmpty)
             DetailsTile(
                 value: vertrag.getNaechsteZahlung()!,
                 description: "nächste Zahlung",
                 lineColor: labelColor),
-          SizedBox(height: 20),
+          const SizedBox(height: 20),
           if ((vertrag.vertragspartner != null) ||
               vertrag.getVertragsBeginn() != null ||
               vertrag.getVertragsEnde() != null ||
               vertrag.getKuendigungsfrist() != null)
-            Text(
+            const Text(
               "Vertragsinformationen",
               textAlign: TextAlign.center,
               style: TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
