+++ frontend/ctfl_vertragsmanager/lib/partials/navbar.dart
@@ -1,4 +1,4 @@
-import 'package:ctfl_vertragsmanager/constants/Color_Themes.dart';
+import 'package:ctfl_vertragsmanager/constants/colors.dart';
 import 'package:flutter/material.dart';
 
 class BaseNavBar extends StatefulWidget implements PreferredSizeWidget {
