+++ frontend/ctfl_vertragsmanager/lib/provider/all_vertraege_provider.dart
@@ -1,10 +1,9 @@
-import 'package:ctfl_vertragsmanager/funktionen/hiveFunctions.dart';
+import 'package:ctfl_vertragsmanager/funktionen/hive_functions.dart';
 import 'package:ctfl_vertragsmanager/models/vertrag.dart';
-import 'package:flutter/material.dart';
-
+import 'package:flutter/foundation.dart';
 import '../models/label.dart';
 
-class all_Vertraege_Provider with ChangeNotifier {
+class AllVertraegeProvider with ChangeNotifier {
   List<Vertrag> _vertraege = [];
 
   List<Vertrag> get vertraege => _vertraege;
@@ -13,25 +12,27 @@ class all_Vertraege_Provider with ChangeNotifier {
     notifyListeners();
   }
 
-  List<Vertrag> get_all_vertraege() {
+  List<Vertrag> getAllVertraege() {
     _vertraege = HiveFunctions.getHiveVertraege().values.toList();
     return _vertraege;
   }
 
-  List<Vertrag> get_all_vertraege_by_Label(Label label) {
-    List<Vertrag> filtered_vertraege =
+  List<Vertrag> getAllVertraegeByLabel(Label label) {
+    List<Vertrag> filteredVertraege =
         _vertraege.where((element) => element.label == label).toList();
-    return filtered_vertraege;
+    return filteredVertraege;
   }
 
-  List<Label> get_all_Labels() {
+  List<Label> getAllLabels() {
     List<Label> labels = [];
     for (Vertrag vertrag in _vertraege) {
       if (!labels.contains(vertrag.label)) {
         labels.add(vertrag.label!);
       }
     }
-    print(labels.length);
+    if (kDebugMode) {
+      print(labels.length);
+    }
     return labels;
   }
 }
