+++ frontend/ctfl_vertragsmanager/lib/provider/cur_vertrag_provider.dart
@@ -1,23 +1,23 @@
-import 'package:ctfl_vertragsmanager/funktionen/hiveFunctions.dart';
+import 'package:ctfl_vertragsmanager/funktionen/hive_functions.dart';
 import 'package:ctfl_vertragsmanager/models/vertrag.dart';
 import 'package:flutter/material.dart';
 
-class cur_Vertrag_Provider with ChangeNotifier {
-  String _cur_Vertrag_id = "-1";
+class CurVertragProvider with ChangeNotifier {
+  String _curVertragId = "-1";
 
-  String get cur_Vertrag_id => _cur_Vertrag_id;
+  String get curVertragId => _curVertragId;
 
-  void set_cur_Vertrag_id(String value) {
-    this._cur_Vertrag_id = value;
+  void setCurVertragId(String value) {
+    _curVertragId = value;
     notifyListeners();
   }
 
-  Future<Vertrag> get_cur_Vertrag() async {
-    return await getHiveVertragById(cur_Vertrag_id);
+  Future<Vertrag> getCurVertrag() async {
+    return await getHiveVertragById(curVertragId);
   }
 
-  void reset_cur_Vertrag_id() {
-    this._cur_Vertrag_id = "-1";
+  void resetCurVertragId() {
+    _curVertragId = "-1";
     notifyListeners();
   }
 }
